+++ lib/src/transaction.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 
 use crate::backend::Timestamp;
 use crate::dag_walk::closest_common_node;
-use crate::index::ReadonlyIndex;
+use crate::default_index_store::ReadonlyIndex;
 use crate::op_store;
 use crate::op_store::OperationMetadata;
 use crate::operation::Operation;
